requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
easyocr>=1.7.0
surya-ocr>=0.6.0
streamlit>=1.30.0
rich>=13.0.0
numpy>=1.24.0
rapidfuzz>=3.0.0
